Refactor: Docummented assembly functions and used local labels
First of all, the bootsector now has oficially become the bootloader V1.0.0. It is not just a simple bootsector now, as it loads data from other sectors such as the kernel.
Many of the functions I made in assembly did not have the ammount of documentation needed and they were mostly vague, leaving any other developers looking at it puzzled because of my bad explanations.
Now all of the assembly functions which are are not part of the main bootloader file have a "header" which provide:
* the function name
* a brief description of what it does and uses to achieve it
* what are the parameters of the function and which register it uses for it
* the return value and the register it is stored in
This should clear any confusion but I also added comments inside of the functions to better understand what it is trying to do.

I also changed to using NASM's Local Labels. No more ugly names starting with '_' and long ones to avoid label collision.
I also switched to using C's predefined "stdint.h" since there is no reason to reinvent the wheel.

Changed the kernel's welcome message to something more, well, welcoming.