Fix: Bootloader now runs on devices with A20 disabled by default
Now, the bootloader can try most methods available to enable the A20 line so that we can access more than 1MB in Protected Mode.
Also, to keep the bootloader under 512bytes, the error messages are now shortened to save precious bytes.
The new format and the error numbers with their meanings can be seen in the README.md