Refactor/Fix: changed tree structure and rewritten makefile
After looking at the linux source code, I found the tree structure to be really easy to understand and I decided to implement it too.
Also, I got tired of looking at the makefile because it was ugly to modify paths and there was also an issue with building the kernel.bin that would make it have an IMMENSE size.
All of that is fixed in this commit, the kernels size being under 3MB.